requirements.txt
# Core ML Libraries
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Monitoring & Visualization
plotly==5.18.0
prometheus-client==0.19.0

# Load Testing
locust==2.18.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1
opencv-python==4.8.1.78

# Model Serialization
joblib==1.3.2
